+++ crates/maple_cli/src/stdio_server/mod.rs
@@ -20,11 +20,7 @@ use once_cell::sync::OnceCell;
 use serde::Serialize;
 use serde_json::json;
 
-use self::providers::{
-    dumb_jump,
-    filer::{self, FilerSession},
-    quickfix, recent_files, BuiltinSession,
-};
+use self::providers::{dumb_jump, quickfix, recent_files};
 use self::rpc::{Call, RpcClient};
 use self::session::{SessionEvent, SessionManager};
 use self::session_client::SessionClient;