+++ crates/maple_cli/src/stdio_server/providers/custom/dumb_jump.rs
@@ -5,6 +5,7 @@ use std::sync::Arc;
 use anyhow::{anyhow, Result};
 use crossbeam_channel::Sender;
 use itertools::Itertools;
+use rayon::prelude::*;
 use serde::Deserialize;
 use serde_json::json;
 
@@ -14,17 +15,42 @@ use crate::dumb_analyzer::{Filtering, RegexSearcher, TagSearcher, Usage, Usages}
 use crate::stdio_server::{
     providers::builtin::OnMoveHandler,
     rpc::Call,
-    session::{EventHandler, NewSession, Session, SessionContext, SessionEvent},
+    session::{
+        note_job_is_finished, register_job_successfully, EventHandle, Session, SessionContext,
+        SessionEvent,
+    },
     write_response, MethodCall,
 };
 use crate::tools::ctags::{get_language, TagsConfig};
 use crate::utils::ExactOrInverseTerms;
 
+/// Internal reprentation of user input.
+#[derive(Debug, Clone, Default)]
+struct SearchInfo {
+    /// Keyword for the tag or regex searching.
+    keyword: String,
+    /// Search terms for further filtering.
+    filtering_terms: ExactOrInverseTerms,
+}
+
+impl SearchInfo {
+    /// Returns true if the filtered results of `self` is a superset of applying `other` on the
+    /// same source.
+    ///
+    /// The rule is as follows:
+    ///
+    /// - the keyword is the same.
+    /// - the new query is a subset of last query.
+    fn has_superset_results(&self, other: &Self) -> bool {
+        self.keyword == other.keyword && self.filtering_terms.contains(&other.filtering_terms)
+    }
+}
+
 fn search_tags(
     dir: &Path,
     extension: &str,
     query: &str,
-    exact_or_inverse_terms: &ExactOrInverseTerms,
+    filtering_terms: &ExactOrInverseTerms,
 ) -> Result<Usages> {
     let ignorecase = query.chars().all(char::is_lowercase);
 
@@ -43,7 +69,7 @@ fn search_tags(
         .search(query, filtering, true)?
         .filter_map(|tag_line| {
             let (line, indices) = tag_line.grep_format(query, ignorecase);
-            exact_or_inverse_terms
+            filtering_terms
                 .check_jump_line((line, indices.unwrap_or_default()))
                 .map(|(line, indices)| Usage::new(line, indices))
         })
@@ -56,14 +82,14 @@ async fn search_regex(
     word: String,
     extension: String,
     cwd: String,
-    exact_or_inverse_terms: &ExactOrInverseTerms,
+    filtering_terms: &ExactOrInverseTerms,
 ) -> Result<Usages> {
     let searcher = RegexSearcher {
         word,
         extension,
         dir: Some(cwd.into()),
     };
-    searcher.search_usages(false, exact_or_inverse_terms).await
+    searcher.search_usages(false, filtering_terms).await
 }
 
 // Returns a combo of tag results and regex results, tag results should be displayed first.
@@ -76,22 +102,23 @@ fn combine(tag_results: Usages, regex_results: Usages) -> Usages {
 }
 
 /// Parses the raw user input and returns the final keyword as well as the constraint terms.
+/// Currently, only one keyword is supported.
 ///
-/// `hel 'fn` => `identifier(s) ++ exact_term/inverse_term`.
+/// `hel 'fn` => `keyword ++ exact_term/inverse_term`.
 ///
 /// # Argument
 ///
-/// - `query`: Initial query user typed in the input window.
-fn parse_raw_input(query: &str) -> (String, ExactOrInverseTerms) {
+/// - `query`: Initial query typed in the input window.
+fn parse_search_info(query: &str) -> SearchInfo {
     let Query {
         exact_terms,
         inverse_terms,
         fuzzy_terms,
     } = Query::from(query);
 
-    // If there is no fuzzy term, use the full query as the identifier,
-    // otherwise restore the fuzzy query as the identifier we are going to search.
-    let (identifier, exact_or_inverse_terms) = if fuzzy_terms.is_empty() {
+    // If there is no fuzzy term, use the full query as the keyword,
+    // otherwise restore the fuzzy query as the keyword we are going to search.
+    let (keyword, filtering_terms) = if fuzzy_terms.is_empty() {
         if !exact_terms.is_empty() {
             (
                 exact_terms[0].word.clone(),
@@ -113,20 +140,25 @@ fn parse_raw_input(query: &str) -> (String, ExactOrInverseTerms) {
         )
     };
 
-    (identifier, exact_or_inverse_terms)
+    SearchInfo {
+        keyword,
+        filtering_terms,
+    }
 }
 
 #[derive(Debug, Clone, Default)]
-pub struct SearchResults {
+struct SearchResults {
     /// Last searching results.
     ///
     /// When passing the line content from Vim to Rust, the performance
     /// of Vim can become very bad because some lines are extremely long,
     /// we cache the last results on Rust to allow passing the line number
     /// from Vim later instead.
-    pub usages: Usages,
-    /// Last query.
-    pub query: String,
+    usages: Usages,
+    /// Last raw query.
+    raw_query: String,
+    /// Last parsed search info.
+    search_info: SearchInfo,
 }
 
 #[derive(Deserialize)]
@@ -148,9 +180,10 @@ enum SearchEngine {
     Both,
 }
 
-async fn handle_dumb_jump_message(
+async fn search_for_usages(
     msg_id: u64,
     params: Params,
+    search_info: Option<SearchInfo>,
     search_engine: SearchEngine,
     force_execute: bool,
 ) -> SearchResults {
@@ -164,44 +197,35 @@ async fn handle_dumb_jump_message(
         return Default::default();
     }
 
-    let (identifier, exact_or_inverse_terms) = parse_raw_input(query.as_ref());
+    let SearchInfo {
+        keyword,
+        filtering_terms,
+    } = search_info.unwrap_or_else(|| parse_search_info(query.as_ref()));
 
     let usages_result = match search_engine {
         SearchEngine::Ctags => {
-            let results = search_tags(
-                Path::new(&cwd),
-                &extension,
-                &identifier,
-                &exact_or_inverse_terms,
-            );
+            let results = search_tags(Path::new(&cwd), &extension, &keyword, &filtering_terms);
             // tags might be incomplete, try the regex way if no results from the tags file.
             let try_regex =
                 results.is_err() || results.as_ref().map(|r| r.is_empty()).unwrap_or(false);
             if try_regex {
-                search_regex(identifier, extension, cwd, &exact_or_inverse_terms).await
+                search_regex(keyword.clone(), extension, cwd, &filtering_terms).await
             } else {
                 results
             }
         }
         SearchEngine::Regex => {
-            search_regex(identifier, extension, cwd, &exact_or_inverse_terms).await
+            search_regex(keyword.clone(), extension, cwd, &filtering_terms).await
         }
         SearchEngine::Both => {
             let tags_future = {
                 let cwd = cwd.clone();
-                let identifier = identifier.clone();
+                let keyword = keyword.clone();
                 let extension = extension.clone();
-                let exact_or_inverse_terms = exact_or_inverse_terms.clone();
-                async move {
-                    search_tags(
-                        Path::new(&cwd),
-                        &extension,
-                        &identifier,
-                        &exact_or_inverse_terms,
-                    )
-                }
+                let filtering_terms = filtering_terms.clone();
+                async move { search_tags(Path::new(&cwd), &extension, &keyword, &filtering_terms) }
             };
-            let regex_future = search_regex(identifier, extension, cwd, &exact_or_inverse_terms);
+            let regex_future = search_regex(keyword.clone(), extension, cwd, &filtering_terms);
 
             let (tags_results, regex_results) =
                 futures::future::join(tags_future, regex_future).await;
@@ -243,26 +267,33 @@ async fn handle_dumb_jump_message(
             (response, Default::default())
         }
     };
+
     write_response(response);
-    SearchResults { usages, query }
+
+    SearchResults {
+        usages,
+        raw_query: query,
+        search_info: SearchInfo {
+            keyword,
+            filtering_terms,
+        },
+    }
 }
 
 #[derive(Debug, Clone, Default)]
-pub struct DumbJumpMessageHandler {
-    /// Last/Latest search results.
-    results: SearchResults,
+pub struct DumbJumpHandle {
+    /// Results from last searching.
+    /// This might be a superset of searching results for the last query.
+    cached_results: SearchResults,
+    /// Current results from refiltering on `cached_results`.
+    current_usages: Option<Usages>,
     /// Whether the tags file has been (re)-created.
     tags_regenerated: Arc<AtomicBool>,
 }
 
-impl DumbJumpMessageHandler {
+impl DumbJumpHandle {
     // TODO: smarter strategy to regenerate the tags?
-    fn regenerate_tags(&mut self, dir: &str, extension: String) {
-        let mut tags_config = TagsConfig::with_dir(dir);
-        if let Some(language) = get_language(&extension) {
-            tags_config.languages(language.into());
-        }
-
+    fn regenerate_tags(&mut self, tags_config: TagsConfig<String>) {
         let tags_searcher = TagSearcher::new(tags_config);
         match tags_searcher.generate_tags() {
             Ok(()) => {
@@ -273,15 +304,60 @@ impl DumbJumpMessageHandler {
             }
         }
     }
+
+    /// Starts a new searching task.
+    async fn start_search(
+        &self,
+        msg_id: u64,
+        params: Params,
+        search_info: SearchInfo,
+    ) -> SearchResults {
+        let job_future = search_for_usages(
+            msg_id,
+            params,
+            Some(search_info),
+            if self.tags_regenerated.load(Ordering::Relaxed) {
+                SearchEngine::Both
+            } else {
+                SearchEngine::Regex
+            },
+            false,
+        );
+
+        tokio::spawn(job_future).await.unwrap_or_else(|e| {
+            tracing::error!(?e, "Failed to spawn task search_for_usages");
+            Default::default()
+        })
+    }
 }
 
 #[async_trait::async_trait]
-impl EventHandler for DumbJumpMessageHandler {
-    async fn handle_on_move(
-        &mut self,
-        msg: MethodCall,
-        context: Arc<SessionContext>,
-    ) -> Result<()> {
+impl EventHandle for DumbJumpHandle {
+    async fn on_create(&mut self, call: Call, context: Arc<SessionContext>) {
+        let (msg_id, params) = parse_msg(call.unwrap_method_call());
+
+        let mut tags_config = TagsConfig::with_dir(params.cwd.clone());
+        if let Some(language) = get_language(&params.extension) {
+            tags_config.languages(language.into());
+        }
+        let job_id = utility::calculate_hash(&tags_config);
+
+        if register_job_successfully(job_id) {
+            tokio::task::spawn_blocking({
+                let mut handler = self.clone();
+                move || {
+                    handler.regenerate_tags(tags_config);
+                    note_job_is_finished(job_id);
+                }
+            });
+        }
+
+        tokio::spawn(async move {
+            search_for_usages(msg_id, params, None, SearchEngine::Regex, true).await;
+        });
+    }
+
+    async fn on_move(&mut self, msg: MethodCall, context: Arc<SessionContext>) -> Result<()> {
         let msg_id = msg.id;
 
         let lnum = msg
@@ -289,7 +365,12 @@ impl EventHandler for DumbJumpMessageHandler {
             .map_err(|_| anyhow!("Missing `lnum` in {:?}", msg))?;
 
         // lnum is 1-indexed
-        if let Some(curline) = self.results.usages.get_line((lnum - 1) as usize) {
+        if let Some(curline) = self
+            .current_usages
+            .as_ref()
+            .unwrap_or(&self.cached_results.usages)
+            .get_line((lnum - 1) as usize)
+        {
             if let Err(error) =
                 OnMoveHandler::create(&msg, &context, Some(curline.into())).map(|x| x.handle())
             {
@@ -301,46 +382,48 @@ impl EventHandler for DumbJumpMessageHandler {
         Ok(())
     }
 
-    async fn handle_on_typed(
-        &mut self,
-        msg: MethodCall,
-        _context: Arc<SessionContext>,
-    ) -> Result<()> {
+    async fn on_typed(&mut self, msg: MethodCall, _context: Arc<SessionContext>) -> Result<()> {
         let (msg_id, params) = parse_msg(msg);
 
-        let job_future = if self.tags_regenerated.load(Ordering::Relaxed) {
-            handle_dumb_jump_message(msg_id, params, SearchEngine::Both, false)
-        } else {
-            handle_dumb_jump_message(msg_id, params, SearchEngine::Regex, false)
-        };
+        let search_info = parse_search_info(&params.query);
 
-        let results = tokio::spawn(job_future).await.unwrap_or_else(|e| {
-            tracing::error!(?e, "Failed to spawn task handle_dumb_jump_message");
-            Default::default()
-        });
-        self.results = results;
+        // Try to refilter the cached results.
+        if self
+            .cached_results
+            .search_info
+            .has_superset_results(&search_info)
+        {
+            let refiltered = self
+                .cached_results
+                .usages
+                .par_iter()
+                .filter_map(|Usage { line, indices }| {
+                    search_info
+                        .filtering_terms
+                        .check_jump_line((line.clone(), indices.clone()))
+                        .map(|(line, indices)| Usage::new(line, indices))
+                })
+                .collect::<Vec<_>>();
+            let total = refiltered.len();
+            let (lines, indices): (Vec<&str>, Vec<&[usize]>) = refiltered
+                .iter()
+                .take(200)
+                .map(|Usage { line, indices }| (line.as_str(), indices.as_slice()))
+                .unzip();
+            let response = json!({
+                "id": msg_id,
+                "provider_id": "dumb_jump",
+                "force_execute": true,
+                "result": { "lines": lines, "indices": indices, "total": total },
+            });
+            write_response(response);
+            self.current_usages.replace(refiltered.into());
+            return Ok(());
+        }
+
+        self.cached_results = self.start_search(msg_id, params, search_info).await;
+        self.current_usages.take();
+
         Ok(())
-    }
-}
-
-pub struct DumbJumpSession;
-
-impl NewSession for DumbJumpSession {
-    fn spawn(call: Call) -> Result<Sender<SessionEvent>> {
-        let mut handler = DumbJumpMessageHandler::default();
-        let (session, session_sender) = Session::new(call.clone(), handler.clone());
-        session.start_event_loop();
-
-        let (msg_id, params) = parse_msg(call.unwrap_method_call());
-        tokio::task::spawn_blocking({
-            let dir = params.cwd.clone();
-            let extension = params.extension.clone();
-            move || handler.regenerate_tags(&dir, extension)
-        });
-        tokio::spawn(async move {
-            handle_dumb_jump_message(msg_id, params, SearchEngine::Regex, true).await;
-        });
-
-        Ok(session_sender)
     }
 }