+++ crates/maple_cli/src/stdio_server/session/mod.rs
@@ -4,9 +4,11 @@ mod manager;
 use std::borrow::Cow;
 use std::collections::HashSet;
 use std::sync::Arc;
+use std::time::Duration;
 
 use anyhow::Result;
 use crossbeam_channel::Sender;
+use futures::Future;
 use once_cell::sync::Lazy;
 use parking_lot::Mutex;
 
@@ -14,22 +16,88 @@ use crate::stdio_server::providers::builtin::on_session_create;
 use crate::stdio_server::{rpc::Call, types::ProviderId, MethodCall};
 
 pub use self::context::{Scale, SessionContext, SyncFilterResults};
-pub use self::manager::{NewSession, SessionManager};
+pub use self::manager::SessionManager;
 
 static BACKGROUND_JOBS: Lazy<Arc<Mutex<HashSet<u64>>>> =
     Lazy::new(|| Arc::new(Mutex::new(HashSet::default())));
 
+pub fn spawn_singleton_job(
+    task_future: impl Future<Output = ()> + Send + Sync + 'static,
+    job_id: u64,
+) {
+    if register_job_successfully(job_id) {
+        tokio::spawn(async move {
+            task_future.await;
+            note_job_is_finished(job_id)
+        });
+    }
+}
+
+pub fn register_job_successfully(job_id: u64) -> bool {
+    let mut background_jobs = BACKGROUND_JOBS.lock();
+    if background_jobs.contains(&job_id) {
+        false
+    } else {
+        background_jobs.insert(job_id);
+        true
+    }
+}
+
+pub fn note_job_is_finished(job_id: u64) {
+    let mut background_jobs = BACKGROUND_JOBS.lock();
+    background_jobs.remove(&job_id);
+}
+
 pub type SessionId = u64;
 
+fn process_source_scale(scale: Scale, context: Arc<SessionContext>) {
+    if let Some(total) = scale.total() {
+        let method = "s:set_total_size";
+        utility::println_json_with_length!(total, method);
+    }
+
+    if let Some(lines) = scale.initial_lines(100) {
+        printer::decorate_lines::<i64>(lines, context.display_winwidth as usize, context.icon)
+            .print_on_session_create();
+    }
+
+    let mut val = context.scale.lock();
+    *val = scale;
+}
+
 #[async_trait::async_trait]
-pub trait EventHandler: Send + Sync + 'static {
-    async fn handle_on_move(&mut self, msg: MethodCall, context: Arc<SessionContext>)
-        -> Result<()>;
-    async fn handle_on_typed(
-        &mut self,
-        msg: MethodCall,
-        context: Arc<SessionContext>,
-    ) -> Result<()>;
+pub trait EventHandle: Send + Sync + 'static {
+    async fn on_create(&mut self, _call: Call, context: Arc<SessionContext>) {
+        const TIMEOUT: Duration = Duration::from_millis(300);
+
+        match tokio::time::timeout(TIMEOUT, on_session_create(context.clone())).await {
+            Ok(scale_result) => match scale_result {
+                Ok(scale) => process_source_scale(scale, context),
+                Err(e) => tracing::error!(?e, "Error occurred on creating session"),
+            },
+            Err(_) => {
+                tracing::debug!(timeout = ?TIMEOUT, "Did not receive value in time");
+                match context.provider_id.as_str() {
+                    "grep" | "grep2" => {
+                        let rg_cmd =
+                            crate::command::grep::RgBaseCommand::new(context.cwd.to_path_buf());
+                        let job_id = utility::calculate_hash(&rg_cmd.inner);
+                        spawn_singleton_job(
+                            async move {
+                                rg_cmd.create_cache().await;
+                            },
+                            job_id,
+                        );
+                    }
+                    _ => {}
+                }
+            }
+        }
+    }
+
+    async fn on_move(&mut self, msg: MethodCall, context: Arc<SessionContext>) -> Result<()>;
+
+    async fn on_typed(&mut self, msg: MethodCall, context: Arc<SessionContext>) -> Result<()>;
 }
 
 #[derive(Debug, Clone)]
@@ -46,7 +114,7 @@ pub struct Session<T> {
 pub enum SessionEvent {
     OnTyped(MethodCall),
     OnMove(MethodCall),
-    Create,
+    Create(Call),
     Terminate,
 }
 
@@ -56,13 +124,13 @@ impl SessionEvent {
         match self {
             Self::OnTyped(msg) => format!("OnTyped, msg_id: {}", msg.id).into(),
             Self::OnMove(msg) => format!("OnMove, msg_id: {}", msg.id).into(),
-            Self::Create => "Create".into(),
+            Self::Create(_) => "Create".into(),
             Self::Terminate => "Terminate".into(),
         }
     }
 }
 
-impl<T: EventHandler> Session<T> {
+impl<T: EventHandle> Session<T> {
     pub fn new(call: Call, event_handler: T) -> (Self, Sender<SessionEvent>) {
         let (session_sender, session_receiver) = crossbeam_channel::unbounded();
 
@@ -92,87 +160,24 @@ impl<T: EventHandler> Session<T> {
         &self.context.provider_id
     }
 
-    fn process_source_scale(&self, scale: Scale) {
-        if let Some(total) = scale.total() {
-            let method = "s:set_total_size";
-            utility::println_json_with_length!(total, method);
-        }
-
-        if let Some(lines) = scale.initial_lines(100) {
-            printer::decorate_lines::<i64>(
-                lines,
-                self.context.display_winwidth as usize,
-                self.context.icon,
-            )
-            .print_on_session_create();
-        }
-
-        let mut val = self.context.scale.lock();
-        *val = scale;
-    }
-
-    async fn handle_create(&mut self) {
-        let context_clone = self.context.clone();
-
-        const TIMEOUT: u64 = 300;
-
-        match tokio::time::timeout(
-            std::time::Duration::from_millis(TIMEOUT),
-            on_session_create(context_clone),
-        )
-        .await
-        {
-            Ok(scale_result) => match scale_result {
-                Ok(scale) => self.process_source_scale(scale),
-                Err(e) => tracing::error!(?e, "Error occurred on creating session"),
-            },
-            Err(_) => {
-                tracing::debug!(timeout = TIMEOUT, "Did not receive value in time");
-                match self.context.provider_id.as_str() {
-                    "grep" | "grep2" => {
-                        let rg_cmd = crate::command::grep::RgBaseCommand::new(
-                            self.context.cwd.to_path_buf(),
-                        );
-                        let job_id = utility::calculate_hash(&rg_cmd.inner);
-                        let mut background_jobs = BACKGROUND_JOBS.lock();
-                        if background_jobs.contains(&job_id) {
-                            tracing::debug!(job_id, "An existing job for grep/grep2");
-                        } else {
-                            tracing::debug!(job_id, "Spawning a background job for grep/grep2");
-                            background_jobs.insert(job_id);
-
-                            tokio::spawn(async move {
-                                let res = rg_cmd.create_cache().await;
-                                let mut background_jobs = BACKGROUND_JOBS.lock();
-                                background_jobs.remove(&job_id);
-                                tracing::debug!(
-                                  job_id,
-                                  result = ?res,
-                                  "The background job is done",
-                                );
-                            });
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-
     async fn process_event(&mut self, event: SessionEvent) -> Result<()> {
         match event {
             SessionEvent::Terminate => self.handle_terminate(),
-            SessionEvent::Create => self.handle_create().await,
+            SessionEvent::Create(call) => {
+                self.event_handler
+                    .on_create(call, self.context.clone())
+                    .await
+            }
             SessionEvent::OnMove(msg) => {
                 self.event_handler
-                    .handle_on_move(msg, self.context.clone())
+                    .on_move(msg, self.context.clone())
                     .await?;
             }
             SessionEvent::OnTyped(msg) => {
                 // TODO: use a buffered channel here, do not process on every
                 // single char change.
                 self.event_handler
-                    .handle_on_typed(msg, self.context.clone())
+                    .on_typed(msg, self.context.clone())
                     .await?;
             }
         }